fix(moryflow): 强化 AI Proxy/积分/发布站点容错并同步文档

+++ apps/moryflow/server/src/ai-proxy/dto/schemas.ts
@@ -156,7 +156,7 @@ export const ChatCompletionRequestSchema = z
 
     // 其他参数
     stop: z.union([z.string(), z.array(z.string())]).optional(),
-    n: z.number().positive().optional(),
+    n: z.number().int().positive().optional(),
     user: z.string().optional(),
   })
   .passthrough(); // 允许额外字段透传
